requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
python-dotenv
mysqlclient
gunicorn
openai
langchain
langchain-core
langchain-openai
langchain-community
langchain-huggingface
langgraph
pymilvus
sentence-transformers

